Added a clean parameter to delete current results.  TODO:  Continue training from existing model if flag not specified.
Prepared input for evaluation script.
Added evaluation script.

+++ segmentation/model1/cifar10_train.py
@@ -124,9 +124,12 @@ def train():
 
 def main(argv=None):  # pylint: disable=unused-argument
   # cifar10.maybe_download_and_extract()
-  if tf.gfile.Exists(FLAGS.train_dir):
-    tf.gfile.DeleteRecursively(FLAGS.train_dir)
-  tf.gfile.MakeDirs(FLAGS.train_dir)
+  if argv[-1] == "clean":
+    if tf.gfile.Exists(FLAGS.train_dir):
+      print("cleaning: " + FLAGS.train_dir)
+      tf.gfile.DeleteRecursively(FLAGS.train_dir)
+    tf.gfile.MakeDirs(FLAGS.train_dir)
+
   train()
 
 if __name__ == '__main__':
